+++ src/multipole.h
@@ -102,7 +102,7 @@ __attribute__((nonnull)) INLINE static void gravity_drift(
               m->m_pole.vel[2] - m->m_pole.min_delta_vel[2]);
 
   const float max_delta_vel =
-      sqrt(dv[0] * dv[0] + dv[1] * dv[1] + dv[2] * dv[2]);
+      sqrtf(dv[0] * dv[0] + dv[1] * dv[1] + dv[2] * dv[2]);
   const float x_diff = max_delta_vel * dt;
 
   /* Conservative change in maximal radius containing all gpart */
@@ -890,7 +890,7 @@ __attribute__((nonnull)) INLINE static void gravity_multipole_compute_power(
   // power[1] += m->M_100 * m->M_100;
 
   // m->power[1] = sqrt(power[1]);
-  m->power[1] = 0.;
+  m->power[1] = 0.f;
 #endif
 #if SELF_GRAVITY_MULTIPOLE_ORDER > 1
   /* 2nd order terms */
